--- src/containers/AdsContainer/index.js
@@ -1,1 +1,0 @@
-export { AdsContainer } from "./AdsContainer";
